KEYCLOAK-9008 CookieStoreRootContextTest stabilization
